Use standard C thread_local storage class specifier
- C11/C17 defines the `_Thread_local` storage class specifier.
- C11/C17 defines the `<threads.h>` header, in which one can find the
  `#define thread_local _Thread_local` macro.
- If the compiler defines the `__STDC_NO_THREADS__` macro, then the
  `<threads.h>` header doesn't exist.
- C23 removes the `_Thread_local` specifier and introduces
  `thread_local` as a keyword.
- The `__thread` storage class specifier is a GCC/clang extension.
- MSVC has support for `thread_local` or `__declspec(thread)`.
- C++11 has the `thread_local` storage class specifier.

Therefore, the nicest option forward is to use `thread_local`. So, if
we're before C23 (so <= C17), we may include `<threads.h>` if it
exists, or define `thread_local` if it is not already defined (out of
caution). If users provide their own definition of `thread_local` in
pre-C23, after including our headers, our definition may conflict with
theirs.