Improve the precision of spin-wait sleeps on Windows
When they are available, use a high-precision timer on Windows
(available since Windows 10 RS4 (1803)).
Those timers require the creation of a HANDLE first but we don't want to
spend time on HANDLE creation and destruction on every spin wait. As
only one OS thread per domain can enter in a spin wait at any given
time, we store the HANDLE in the domain state, so that it can be created
and destroyed along with the domain.

Co-authored-by: Samuel Hym <samuel.hym@rustyne.lautre.net>