Remove dependency on platform.h from io.h
WIP: tweaks the channel structure to be the same as in 4.x with a void *
for the mutex. This is naff, though, because we always have a mutex.
Better would be to do the required incantations such that a buffer of
the appropriate size is included in the typedef (then, in
post-winpthreads times, this would be sizeof(HANDLE) for Win32).