Use `target` instead of `host` to detect the C toolchain
Recall that the only currently officially supported configurations are
when `build` ~ `host` = `target`, where '~' means that the code
generated for `host` runs on `build` even when they differ (such as when
`build` is `x86_64-pc-cygwin` and `host` is `x86_64-pc-windows` (MSVC)
or `x86_64-w64-mingw32`).

Even though this is not officially supported yet, many projects (such as
Mirage and the opam-cross-* repositories) use OCaml cross compilers. All
those projects generate a cross compiler by assuming a non-cross OCaml
compiler is available in `PATH` (where non-cross means generating code
that will run on `host`). For that they need a C compiler and binutils
for `target` in order to build the runtime. (Note that the non-cross
compiler will link its own (`build`/`host`) runtime into the generated
`.opt` cross compilers rather than the just-compiled target runtime.)

Indeed, in that setup the runtime will run (only) on the `target` so:

- set `cross_compiling` by comparing `build` to `target` (rather than to
  `host`), as this variable will be used later
- use `target` to set up the tool prefix,
- as the libtool configuration will configure a `build` to `host`
  toolchain, temporarily assign `host*` values to `target*` values.

Note that all these changes are transparent when `host` = `target`.