Use `target` instead of `host` to detect the C toolchain
Recall that the only currently officially supported configurations are
when `build` ~ `host` = `target`, where '~' means that the code
generated for `host` runs on `build` even when they differ (such as when
`build` is `x86_64-pc-cygwin` and `host` is `x86_64-pc-windows` (MSVC)
or `x86_64-w64-mingw32`).

Still, many projects use OCaml cross compilers. All those projects
generate a cross compiler by assuming a non-cross OCaml compiler is
available in `PATH` (where non-cross means generating code that will run
on `host`). For the cross compiler, a C compiler and binutils for
`target` are necessary to build the target runtime. (Note that the
non-cross compiler will link its own (`build`/`host`) runtime into the
generated `.opt` cross compilers rather than the just-compiled target
runtime.)

In that setup the runtime that will be compiled to create a cross
compiler will run only on the `target` so this commit:

- sets `cross_compiling` by comparing `build` to `target` (rather than
  to `host`), as this variable will be used later,
- uses `target` to set up the tool prefix,
- temporarily assigns `host*` values to `target*` values during the
  libtool configuration, as this detects a `build` to `host` toolchain.

Note that all these changes are transparent when `host` = `target`.