Allow `ocaml` in a triplet target to build freestanding cross compilers
Allow `ocaml` to be used as the last component of the target triplet in
case we are using a custom toolchain for a freestanding target. The
target triplet is then temporarily rewritten to "<arch>-none" to compute
the canonical target.

This allows to use a `*-*-ocaml-` prefixes (ie `x86_64-solo5-ocaml-`) to
create cross-compiler toolchains dedicated to specific freestanding
targets