Allow using clang-cl with *-pc-windows hosts
clang-cl is a C/C++ compiler that is ABI and command-line compatible
with MSVC. The executable is named clang-cl. Hard-coding CC=cl
prevents selecting clang-cl at configure-time.

The configure script is *not* smart enough to select cl when the host
is *-pc-windows and another cc is installed (e.g., mingw or cygwin).

Fallback to CC=cl if the user hasn't specified a CC at configure-time.