Fix static builds of the compiler
ocamlc -custom needs to include <caml/mlvalues.h> and therefore the
build has -I runtime in order to facilitate this. -I runtime breaks
cross-compilation, since it causes the .a files in runtime to be found
by the host ocamlopt instead of its own.

This workaround defeats that problem by adding + to VPATH as the first
include, which will cause the host compiler's Standard Library directory
(which includes the runtime .a files) to be included first in the search
path. When building the compiler's libraries (which should be using the
target's Standard Library, not the host's), -nostdlib -I stdlib is
already inserted before VPATH by the build system, which means this
mechanism never interferes with the selection of the correct Standard
Library.